pal: remove implicit routing to SCO when BT_SCO=on is received
There might be a scenario where BT_SCO=on is sent but SCO is not set
as preferred device at APM due to SCO getting disconnected immediately.
Implicit routing of streams to SCO during SCO on handling will cause
streams to stay with SCO even if SCO is not connected.

Fixed this by removing implicit routing of stream to SCO device.

Change-Id: I9a95e7d70582067c3759d1826d15d4e6ae076428